stomped on vc2k5 warning (size_t -> uint)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@1191 3db68df2-c116-0410-a063-a993310a9797

+++ source/scripting/ScriptingHost.cpp
@@ -323,7 +323,7 @@ CStrW ScriptingHost::ValueToUCString( const jsval value )
 		throw PSERROR_Scripting_ConversionFailed();
 
 	jschar *strptr=JS_GetStringChars(string);
-	uint length=JS_GetStringLength(string);
+	size_t length=JS_GetStringLength(string);
 	return CStrW(std::wstring(strptr, strptr+length));
 }
